refactor fuelmap generation
