+++ src/pyrolib/fuelmap/__init__.py
@@ -1,0 +1,15 @@
+from .fuels import (
+    BalbiFuel,
+    show_fuel_classes,
+)
+from .fuelmap import (
+    FuelMap,
+)
+from .scenario import (
+    Scenario,
+    show_default_scenario,
+)
+from .utility import (
+    fire_array_2d_to_3d,
+    fire_array_3d_to_2d,
+)
